+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandListener.java
@@ -1,17 +1,10 @@
 package com.github.johnnyjayjay.discord.commandapi;
 
-import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.Permission;
-import net.dv8tion.jda.core.entities.Member;
-import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
 
 import java.util.HashMap;
 import java.util.Map;
-
-import static java.lang.String.join;
-import static java.lang.String.format;
 
 
 class CommandListener extends ListenerAdapter {
@@ -43,25 +36,8 @@ class CommandListener extends ListenerAdapter {
                 if (cmd.getExecutor() != null) {
                     cmd.getExecutor().onCommand(new CommandEvent(event.getJDA(), event.getResponseNumber(), event.getMessage(), cmd, settings),
                             event.getMember(), event.getChannel(), cmd.getArgs());
-                } else if (settings.getHelpLabels().contains(cmd.getLabel()) && event.getGuild().getSelfMember().hasPermission(event.getChannel(), Permission.MESSAGE_WRITE)) {
-                    this.sendInfo(event.getMember(), event.getChannel(), prefix, cmd.getArgs());
                 }
             }
-        }
-    }
-
-    private void sendInfo(Member member, TextChannel channel, String prefix, String[] args) {
-        EmbedBuilder builder = new EmbedBuilder().setTitle("Help").setColor(settings.getHelpColor());
-        if (args.length == 0) {
-            String helpLabels = format("[%s]", join("|", settings.getHelpLabels()));
-            builder.appendDescription(format("To learn more about a specific command, just call `%s%s <label>`.\n", prefix, helpLabels))
-                    .appendDescription("The following commands are currently available:\n");
-            builder.addField("Commands", format("```\n%s%s```", prefix, join(format(", %s", prefix), settings.getCommands().keySet())), false);
-            channel.sendMessage(builder.build()).queue();
-        } else if (args.length == 1 && settings.getCommands().containsKey(args[0])) {
-            builder.appendDescription(format("**Command Info for:** `%s`\n\n", args[0]))
-                    .appendDescription(settings.getCommands().get(args[0]).info(member));
-            channel.sendMessage(builder.build()).queue();
         }
     }
 }