+++ changelog.md
@@ -1,9 +1,13 @@
 # Changelog
 
-### Help Command changes
+### 3.2
 - Added classes AbstractHelpCommand and DefaultHelpCommand
 - Added method `getCommandSettings` in `CommandEvent`
 - Annotated `ICommand` as a `FunctionalInterface`
+- Removed supported for help label methods in `CommandSettings`
+- Cleaned `CommandListener`
+- Deprecated method `String info(Member)` in `ICommand`;
+    New method is `Message info(Member, String, Set<String>)` because it is more flexible.
 
 ### 3.1_1
 - Updated to JDA version 3.7.1_387